refactor: moves unstruct funcs to its own pkg
Operations on unstructured k8s objects should be kept in its own pkg for
better code organization. We use this struct to dynamically handle
watched CRs as we do not want to import types of every component for which
we provide platform capabilites. Often such types might not exist as
as a dedicated client-go library with generated stubs.

This change creates new `pkg/unstruct` where we should start organizing
helper functions.
