+++ pkg/unstruct/funcs.go
@@ -1,4 +1,4 @@
-package cluster
+package unstruct
 
 import (
 	"context"
